Added new DrawTextureBuffer
The new feature is disabled by default and can be turend on by calling.
    poro::IPlatform::Instance()->GetGraphics()->SetDrawTextureBuffering(true);

The DrawTextureBuffer works so that consecutive calls to DrawTexture are buffered.
The buffer is automatically flushed using a single draw call when one of the following things happen:
 - The used texture changes.
 - The used color changes.
 - The used blendmode changes.
 - EndRendering() is called.
 - Any other Draw function except the two DrawTexture() functions are called.
 - The buffer is full.
 - Turinging of the buffering.

The performance implication of this buffer depends on the implementation but in a case where you draw alot of stuff from one texture. i.e.
a texture atlas there could be a big performance boost.

- Denamo