Add signed/unsigned char operator>>
Should fix certain fields on CellData receiving junk memory when
attribute exists in the XML.

Because there's no operator>> for these field types, it selects the
templated version of the function which does nothing.
https://github.com/gummikana/poro/blob/7d48a2043cf19bc75dc9e650d2a6a1a732d8cbd6/source/utils/xml/cvarreference.h#L118-L124

Since it calls a function that does nothing, CAnyContainerCast in the
type conversion case returns whatever data happened to be on the stack,
leading to these fields receiving a junk value.

For CellData, it looks like this junk value is consistently 5.

This issue affects the following fields:
    - gas_speed,
    - gas_upwards_speed,
    - gas_horizontal_speed,
    - gas_downwards_speed,
    - and liquid_sprite_stains_check_offset,
    - maybe more, I don't know all fields/types here.

Of these fields, only liquid_sprite_stains_check_offset is used by the
vanilla game. The gas fields are always kept at their default value so
it's not an issue for the base game but it is a problem for modders.